Update events.md
Edits to improve readability. Added some missing items to the table of contents. Reworded some section titles and changed the title pattern from title case to sentence case (in general, the more capital letters, the harder a passage is to read). Added a definition of PMC at the first instance of use. Added links from section references to those sections.