Update index.md
More minor text edits. 
The short lines in source code seem to be an artifact of the old CMS system editor. There is no need to preserve them, especially when the line break happens in the middle of a link statement or similar element and can conceal a formatting issue.