Use more appropriative language in about name
I think the About Name section of the website could use more colonial and appropriative language. There is so much history with The Apache Software Foundation. Much more than any of the Apache peoples. I mean, ASF has gone through so much history in the past 20 or so years. They should be proud of appropriating the name of a people that, in the language of the founder, "succumbed to the invasion from the West." It just sounds so much better when you call yourselves Apache. You all are the real warriors.