Correct the link for new PMC member to `newpmcmember` from `newpmc`
There were a few documentation links (in comdev site mainly) that
lined to #newpmc anchor, and since the links were shown
verbatim in the page, they could be misunderstood - it was not about
new PMC but about new PMC menber, and since we are now reminding
projects that they should not invite "new PMC" but "new PMC member",
correcting it in the link is a good thing to do.

Separate PR in comdev site corrects those links (and changes them
to be replaced with "New PMC member" name.