New marketing category and add "creator" FAQ (#390)
* New marketing category and add "creator" FAQ

The often asked question "Can I refer to an individual, organization or community as 'founding' an ASF project?" is currently on the Newsroom page. Because this is one of the more common brand violations that occur, it would make more sense to appear in the FAQ vs. Newsroom. I'm proposing the addition to the FAQ, and will be proposing the deletion (and re-routing) to the Newsroom page next.

* Update faq.md

Updating wording to my branch based on feedback