Add files via upload (#578)
Added the new logo to the emeritus-request form, using the current version of Open Office. Please confirm that the form appears as it should, and that you can open the ODT file in your version of Open Office. If this works, I will similarly update and commit the other forms under discussion.

This does not address the many other changes that have been suggested in the email thread. I want to get this correct first, before moving ahead.